chore: rename preview commands
