+++ packages/plugin-core/src/commands/index.ts
@@ -42,8 +42,8 @@ import { SetupWorkspaceCommand } from "./SetupWorkspace";
 import { ShowHelpCommand } from "./ShowHelp";
 import { ShowNoteGraphCommand } from "./ShowNoteGraph";
 import { ShowSchemaGraphCommand } from "./ShowSchemaGraph";
+import { ShowLegacyPreviewCommand } from "./ShowLegacyPreview";
 import { ShowPreviewCommand } from "./ShowPreview";
-import { ShowPreviewV2Command } from "./ShowPreviewV2";
 import { SignInCommand } from "./SignIn";
 import { SignUpCommand } from "./SignUp";
 import { SiteBuildCommand } from "./SiteBuild";
@@ -113,8 +113,8 @@ const ALL_COMMANDS = [
   ShowHelpCommand,
   ShowNoteGraphCommand,
   ShowSchemaGraphCommand,
+  ShowLegacyPreviewCommand,
   ShowPreviewCommand,
-  ShowPreviewV2Command,
   SignInCommand,
   SignUpCommand,
   SiteBuildCommand,
